+++ src/rules/__tests__/a11y-tooltip-interactive-trigger.test.js
@@ -51,7 +51,22 @@ ruleTester.run('non-interactive-tooltip-trigger', rule, {
     `import {Tooltip, Link} from '@primer/react';
     <Tooltip aria-label="Supplementary text" direction="e">
       <Link href="https://github.com">Link</Link>
-    </Tooltip>`
+    </Tooltip>`,
+    `
+    import {Tooltip, Link} from '@primer/react';
+    <Tooltip aria-label={avatar.avatarName} direction="e">
+      <Link href={avatar.avatarLink} underline={true}>
+        User avatar
+      </Link>
+    </Tooltip>` ,
+    `
+    import {Tooltip, Link} from '@primer/react';
+    <Tooltip aria-label="product" direction="e">
+      <Link href={productLink}>
+        Product
+      </Link>
+    </Tooltip>
+    `
   ],
   invalid: [
     {
@@ -96,7 +111,7 @@ ruleTester.run('non-interactive-tooltip-trigger', rule, {
       </Tooltip>`,
       errors: [
         {
-          messageId: 'anchorTagWithoutHref'
+          messageId: 'nonInteractiveLink'
         }
       ]
     },
@@ -108,7 +123,7 @@ ruleTester.run('non-interactive-tooltip-trigger', rule, {
       </Tooltip>`,
       errors: [
         {
-          messageId: 'anchorTagWithoutHref'
+          messageId: 'nonInteractiveLink'
         }
       ]
     },
@@ -120,7 +135,7 @@ ruleTester.run('non-interactive-tooltip-trigger', rule, {
       </Tooltip>`,
       errors: [
         {
-          messageId: 'hiddenInput'
+          messageId: 'nonInteractiveInput'
         }
       ]
     },
@@ -132,7 +147,43 @@ ruleTester.run('non-interactive-tooltip-trigger', rule, {
       </Tooltip>`,
       errors: [
         {
-          messageId: 'hiddenInput'
+          messageId: 'nonInteractiveInput'
+        }
+      ]
+    },
+    {
+      code: `
+      import {Tooltip, Button} from '@primer/react';
+      <Tooltip aria-label="Supplementary text" direction="e">
+        <Button disabled>Save</Button>
+      </Tooltip>`,
+      errors: [
+        {
+          messageId: 'nonInteractiveTrigger'
+        }
+      ]
+    },
+    {
+      code: `
+      import {Tooltip, Button} from '@primer/react';
+      <Tooltip aria-label="Supplementary text" direction="e">
+        <IconButton disabled>Save</IconButton>
+      </Tooltip>`,
+      errors: [
+        {
+          messageId: 'nonInteractiveTrigger'
+        }
+      ]
+    },
+    {
+      code: `
+      import {Tooltip, Button} from '@primer/react';
+      <Tooltip aria-label="Supplementary text" direction="e">
+        <input disabled>Save</input>
+      </Tooltip>`,
+      errors: [
+        {
+          messageId: 'nonInteractiveInput'
         }
       ]
     },
@@ -159,7 +210,7 @@ ruleTester.run('non-interactive-tooltip-trigger', rule, {
       </Tooltip>`,
       errors: [
         {
-          messageId: 'anchorTagWithoutHref'
+          messageId: 'nonInteractiveLink'
         }
       ]
     }